requirements.txt
fastmcp>=0.4.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
mcp>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
